Wrap reload config in straight transaction
Finally, a note on the use of transactions for optimizations. There are a number of setup and tear-down operations associated with package management. For example, to keep track of when packages need to be rebuilt, straight.el keeps a persistent build cache. Normally, this cache needs to be read and written after every package install. But that is very slow: much better is to load it at the first package install, and to save it at the last package install. The question then is how to identify the last package install. This is not possible in general (although in the special case of initial Emacs init, after-init-hook can be used), so straight.el falls back on the transaction system. By wrapping the entire operation in a transaction, straight.el can safely optimize the loading and saving of the build cache, significantly improving performance. For this reason, reloading your init-file is likely to be rather slow if you do not wrap the call in a transaction using straight-transaction.