Refine my/magit-search-git-log-for-change to handle multiple lines
If region spans multiple lines does regex or of each trimmed line.
This effectively returns all changes to that set of lines. Or anything
in the file that matches that one of those lines.

Lines containing common patterns that appear throughout the file can
lead to unrelated results. For example '(interactive)' or ':straight
nil' in init.el would lead to a large number of unrelated results as
they a very common occurrence.

If this becomes a problem these common lines could be filtered.