Update topiary smart yank
Like 'yank'. But calling 'yank' again will call 'yank-pop'.

This means we can cycle through the kill ring without using M-y. The
reasoning is I almost never paste the same things multiple times in a
row. So it's more useful to be able to perform yank pop on subsequent
invocations.