Fix indent region deleting code and breaking syntax tree when code folded
This worked fine in emacs 28 but not in emacs 29.

Repro code:

(defun foo ()
  (+
   1000))

(defun bar ()
  (+
   1000))

Fold the code above and place the cursor at the end of the foo expression.
Run indent-region and it will break the syntax tree.

The fix was to remove:

(overlay-put ov 'display "...")

This seems redundant. My best guess was it was breaking the hideshow
overlay by setting the 'display property again. Maybe?